+++ packages/@ionic/cli/src/commands/package/build.ts
@@ -22,7 +22,16 @@ const ANDROID_BUILD_TYPES = ['debug', 'release'];
 const IOS_BUILD_TYPES = ['development', 'ad-hoc', 'app-store', 'enterprise'];
 const APP_STORE_COMPATIBLE_TYPES = ['release', 'app-store', 'enterprise'];
 const BUILD_TYPES = ANDROID_BUILD_TYPES.concat(IOS_BUILD_TYPES);
+const ANDROID_ARTIFACT_TYPES = ['aab', 'apk'];
+const IOS_ARTIFACT_TYPES = ['ipa', 'dsym'];
+const ARTIFACT_TYPES = ANDROID_ARTIFACT_TYPES.concat(IOS_ARTIFACT_TYPES);
 const TARGET_PLATFORM = ['Android', 'iOS - Xcode 11 (Preferred)', 'iOS - Xcode 10'];
+
+interface Artifact {
+  name: string;
+  url: string;
+  artifact_type: string;
+}
 
 export interface PackageBuild {
   job_id: number;
@@ -45,6 +54,7 @@ export interface PackageBuild {
   distribution_credential_name: string;
   job: any;
   distribution_trace: string;
+  artifacts: Artifact[];
 }
 
 interface DownloadUrl {
@@ -74,6 +84,12 @@ Customizing the build:
 Deploying the build to an App Store:
 - The ${input('--destination')} option can be used to deliver the app created by the build to the configured App Store. \
 This can be used only together with build type ${input('release')} for Android and build types ${input('app-store')} or ${input('enterprise')} for iOS.
+
+Downloading build artifacts:
+- By default once the build is complete, all artifacts are downloaded for the selected platform. ${input('aab')} and ${input('apk')} for Android \
+${input('ipa')} and ${input('dsym')} for iOS.
+- The ${input('--artifact-type')} option can be used to limit artifact downloads to only of that type. For instance, with Android, you can specify ${input('aab')} \
+if you do not wish to download ${input('apk')}.
 `,
       footnotes: [
         {
@@ -87,8 +103,10 @@ This can be used only together with build type ${input('release')} for Android a
         'android debug --environment="My Custom Environment Name"',
         'android debug --native-config="My Custom Native Config Name"',
         'android debug --commit=2345cd3305a1cf94de34e93b73a932f25baac77c',
+        'android debug --artifact-type=aab',
+        'android debug --aab-name="my-app-prod.aab" --apk-name="my-app-prod.apk"',
         'ios development --signing-certificate="iOS Signing Certificate Name" --build-stack="iOS - Xcode 9"',
-        'ios development --signing-certificate="iOS Signing Certificate Name" --build-file-name=my_custom_file_name.ipa',
+        'ios development --signing-certificate="iOS Signing Certificate Name" --ipa-name=my_custom_file_name.ipa',
         'ios app-store --signing-certificate="iOS Signing Certificate Name" --destination="Apple App Store Destination"',
       ],
       inputs: [
@@ -146,7 +164,41 @@ This can be used only together with build type ${input('release')} for Android a
           name: 'build-file-name',
           summary: 'The name for the downloaded build file',
           type: String,
+          groups: [MetadataGroup.DEPRECATED],
+          spec: { value: 'name' },
+        },
+        {
+          name: 'ipa-name',
+          summary: 'The name for the downloaded ipa file',
+          type: String,
           groups: [MetadataGroup.ADVANCED],
+          spec: { value: 'name' },
+        },
+        {
+          name: 'dsym-name',
+          summary: 'The name for the downloaded dsym file',
+          type: String,
+          groups: [MetadataGroup.ADVANCED],
+          spec: { value: 'name' },
+        },
+        {
+          name: 'apk-name',
+          summary: 'The name for the downloaded apk file',
+          type: String,
+          groups: [MetadataGroup.ADVANCED],
+          spec: { value: 'name' },
+        },
+        {
+          name: 'aab-name',
+          summary: 'The name for the downloaded aab file',
+          type: String,
+          groups: [MetadataGroup.ADVANCED],
+          spec: { value: 'name' },
+        },
+        {
+          name: 'artifact-type',
+          summary: `The artifact type (${ARTIFACT_TYPES.map(v => input(v)).join(', ')})`,
+          type: String,
           spec: { value: 'name' },
         },
       ],
@@ -169,14 +221,14 @@ This can be used only together with build type ${input('release')} for Android a
     const buildTypes = inputs[0] === 'ios' ? IOS_BUILD_TYPES : ANDROID_BUILD_TYPES;
 
     // validate that the build type is valid for the platform
-    let reenterBuilType = false;
+    let reenterBuildType = false;
     if (inputs[1] && !buildTypes.includes(inputs[1])) {
-      reenterBuilType = true;
+      reenterBuildType = true;
       this.env.log.nl();
       this.env.log.warn(`Build type ${strong(inputs[1])} incompatible for ${strong(inputs[0])}; please choose a correct one`);
     }
 
-    if (!inputs[1] || reenterBuilType) {
+    if (!inputs[1] || reenterBuildType) {
       const typeInput = await this.env.prompt({
         type: 'list',
         name: 'type',
@@ -218,6 +270,26 @@ This can be used only together with build type ${input('release')} for Android a
     if (options['destination'] && !APP_STORE_COMPATIBLE_TYPES.includes(inputs[1])) {
       throw new FatalException(`Build with type ${strong(String(inputs[1]))} cannot be deployed to App Store`);
     }
+
+    if (options['artifact-type'] && (Array.isArray(options['artifact-type']) || typeof options['artifact-type'] === 'string')) {
+      const artifactTypes = Array.isArray(options['artifact-type']) ? options['artifact-type'] : [options['artifact-type']];
+      let unsupported: string[];
+
+      switch (inputs[0]) {
+        case 'android':
+          unsupported = artifactTypes.filter(type => !ANDROID_ARTIFACT_TYPES.includes(type));
+          break;
+        case 'ios':
+          unsupported = artifactTypes.filter(type => !IOS_ARTIFACT_TYPES.includes(type));
+          break;
+        default:
+          throw new FatalException(`Unsupported platform ${inputs[0]}`);
+      }
+
+      if (unsupported.length) {
+        throw new FatalException(`Unsupported artifact types for platform ${inputs[0]}: ${[...unsupported]}`);
+      }
+    }
   }
 
   async run(inputs: CommandLineInputs, options: CommandLineOptions): Promise<void> {
@@ -237,20 +309,13 @@ This can be used only together with build type ${input('release')} for Android a
     let build = await this.createPackageBuild(appflowId, token, platform, buildType, options);
     const buildId = build.job_id;
 
-    let customBuildFileName = '';
-    if (options['build-file-name']) {
-      if (typeof (options['build-file-name']) !== 'string' || !fileUtils.isValidFileName(options['build-file-name'])) {
-        throw new FatalException(`${strong(String(options['build-file-name']))} is not a valid file name`);
-      }
-      customBuildFileName = String(options['build-file-name']);
-    }
-
     const details = columnar([
       ['App ID', strong(appflowId)],
       ['Build ID', strong(buildId.toString())],
       ['Commit', strong(`${build.commit.sha.substring(0, 6)} ${build.commit.note}`)],
       ['Target Platform', strong(build.stack.friendly_name)],
       ['Build Type', strong(build.build_type)],
+      ['Artifact Type(s)', options['artifact-type'] ? strong(`${options['artifact-type']}`.toUpperCase()) : weak('all available')],
       ['Security Profile', build.profile_tag ? strong(build.profile_tag) : weak('not set')],
       ['Environment', build.environment_name ? strong(build.environment_name) : weak('not set')],
       ['Native Config', build.native_config_name ? strong(build.native_config_name) : weak('not set')],
@@ -267,13 +332,42 @@ This can be used only together with build type ${input('release')} for Android a
       throw new Error(`Build ${build.state}`);
     }
 
-    const url = await this.getDownloadUrl(appflowId, buildId, token);
-    if (!url.url) {
-      throw new Error('Missing URL in response');
+    const availableArtifactTypes = build.artifacts.map((artifact) => artifact.artifact_type);
+
+    if (Array.isArray(options['artifact-type'])) {
+      let errors = [];
+      for (const artifactType of new Set(options['artifact-type'])) {
+        try {
+          await this.downloadArtifact(appflowId, buildId, artifactType, token, options);
+        } catch (error) {
+          errors.push(error);
+        }
+      }
+
+      if (errors.length) {
+        throw new FatalException(`There were issues downloading artifacts: ${errors.join('\n')}`);
+      }
+
+    } else if (typeof options['artifact-type'] == 'string') {
+      await this.downloadArtifact(appflowId, buildId, options['artifact-type'], token, options);
+    } else {
+      for (const artifactType of availableArtifactTypes) {
+        await this.downloadArtifact(appflowId, buildId, artifactType.toUpperCase(), token, options);
+      }
+    }
+  }
+
+  async sanitizeString(value: string | string[] | boolean | null | undefined): Promise<string> {
+
+    if (!value || typeof (value) !== 'string') {
+      return '';
     }
 
-    const filename = await this.downloadBuild(url.url, customBuildFileName);
-    this.env.log.ok(`Build completed: ${filename}`);
+    if (!fileUtils.isValidFileName(value)) {
+      throw new FatalException(`${strong(String(value))} is not a valid file name`);
+    }
+
+    return String(value);
   }
 
   async createPackageBuild(appflowId: string, token: string, platform: string, buildType: string, options: CommandLineOptions): Promise<PackageBuild> {
@@ -325,8 +419,8 @@ This can be used only together with build type ${input('release')} for Android a
     }
   }
 
-  async getDownloadUrl(appflowId: string, buildId: number, token: string): Promise<DownloadUrl> {
-    const { req } = await this.env.client.make('GET', `/apps/${appflowId}/packages/${buildId}/download`);
+  async getDownloadUrl(appflowId: string, buildId: number, artifactType: string, token: string): Promise<DownloadUrl> {
+    const { req } = await this.env.client.make('GET', `/apps/${appflowId}/packages/${buildId}/download?artifact_type=${artifactType}`);
     req.set('Authorization', `Bearer ${token}`).send();
 
     try {
@@ -397,5 +491,26 @@ This can be used only together with build type ${input('release')} for Android a
     fs.copyFileSync(tmpFile, filename);
     fs.unlinkSync(tmpFile);
     return filename;
+  }
+
+  async downloadArtifact(appflowId: string, buildId: number, artifactType: string, token: string, options: CommandLineOptions) {
+
+    const url = await this.getDownloadUrl(appflowId, buildId, artifactType.toUpperCase(), token);
+
+    if (!url.url) {
+      throw new Error(`Artifact type '${artifactType}' not found`);
+    }
+
+    let customBuildFileName = '';
+
+    if (options['build-file-name']) {
+      this.env.log.warn(`The ${input('--build-file-name')} option has been deprecated. Please use ${input('--ipa-name')}, ${input('--apk-name')} or ${input('--{ artifact }-name')}.`);
+      customBuildFileName = await this.sanitizeString(options['build-file-name']);
+    } else {
+      customBuildFileName = await this.sanitizeString(options[`${artifactType.toLowerCase()}-name`]);
+    }
+
+    const filename = await this.downloadBuild(url.url, customBuildFileName);
+    this.env.log.ok(`Artifact downloaded: ${filename}`);
   }
 }